requirements.txt
Django==5.0.3
djangorestframework==3.15.1
requests==2.31.0
beautifulsoup4==4.12.3
